% Generated by roxygen2 (4.0.0): do not edit by hand
\name{source}
\alias{source}
\title{Overwrite built-in source function.}
\usage{
source(file, local = FALSE, echo = verbose, print.eval = echo,
  verbose = getOption("verbose"), prompt.echo = getOption("prompt"),
  max.deparse.length = 150, chdir = FALSE,
  encoding = getOption("encoding"), continue.echo = getOption("continue"),
  skip.echo = 0, keep.source = getOption("keep.source"))
}
\description{
Overwrite built-in source function.
}

