% Generated by roxygen2 (4.0.0): do not edit by hand
\name{syberia_objects}
\alias{syberia_objects}
\title{Find all the syberia objects of the given type and subtype in a Syberia project.}
\usage{
syberia_objects(pattern = "", type = NULL, subtype = NULL,
  root = syberia_root(), by_mtime = TRUE, fixed = FALSE)
}
\arguments{
\item{pattern}{character. A set of characters by which to filter.
This uses the same format as the popular ctrl+p plugin for vim.
Namely, it will look for adjacent instances of such characters
regardless of any interpolating characters. For example,
'ace' will match 'abcde' but also 'abcdfghe' but not 'aebcd'.}

\item{type}{character. A subdirectory to look in. For example,
\code{type = 'models'} will look in the \code{models} subdirectory
of the root directory of the Syberia project.}

\item{subtype}{character. A subdirectory to look in inside of the
subdirectory given by \code{type}. For example,
\code{type = 'models', subtype = 'dev'} will look in the \code{dev}
subdirectory of the \code{models} subdirectory of the root directory
of the Syberia project.}

\item{root}{character. The root of the syberia project. The default
is \code{syberia_root()}.}

\item{by_mtime}{logical. Whether or not to sort the models in descending
order by last modified time. The default is \code{TRUE}.}

\item{fixed}{logical. Whether or not to use smart interpolation, like in
the description for the \code{pattern} argument. If \code{TRUE},
only substring matching is used.}
}
\value{
a list of filenames containing syberia objects
}
\description{
Syberia objects can refer to models, data sources, or tests. The essence
of the idea is that the \code{pattern} parameter specifies a set of consecutive
character by which to look for, and \code{type} specifies the subdirectory
(an additional subdirectory of that directory can be set using the \code{subtype}
parameter).
}
\details{
For example, if we are looking for models in the prod environment matching "gbm",
we could try: \code{syberia_objects('gbm', 'models', 'prod')}.

Note, however, that the first argument (\code{pattern}) does not look for a
substring match, but an interpolated match: for example, looking for 'abc'
will match "a1b2c" or "model_a/submodel_bc" but will not match "acb" or
any string where the characters 'a', 'b', and 'c' do not appear consecutively
(with arbitrary strings in between them).
}
\seealso{
\code{\link{syberia_models}}
}

