% Generated by roxygen2 (4.0.0): do not edit by hand
\name{syberia_resource_with_modification_tracking}
\alias{syberia_resource_with_modification_tracking}
\title{Fetch a syberia resource with modification tracking.}
\usage{
syberia_resource_with_modification_tracking(...)
}
\arguments{
\item{...}{same as arguments to \code{syberia_resource}}
}
\description{
Modification tracking refers to determining whether the resource has
changed since its last execution by syberia. This is accomplished
using the output of \code{syberia_resource}, which will return
a list that includes modification time of the last time syberia
encountered the resource. If a modification is spotted, the
Syberia cache entry `runtime/any_modified` will be set to \code{TRUE}.
}
\seealso{
\code{\link{syberia_resource}}
}

